+++ packages/core/src/plugins/resource-timeline/Expander.svelte
@@ -1,24 +1,25 @@
 <script>
     import {getContext} from 'svelte';
-    import {getPayload, isFunction, toViewWithLocalDates} from '#lib';
+    import {contentFrom, getPayload, isFunction, toViewWithLocalDates} from '#lib';
 
     let {resource} = $props();
 
-    let {resources, view, options: {resourceExpand, theme}} = $derived(getContext('state'));
+    let {resources, view, options: {buttonText, icons, resourceExpand, theme}} = $derived(getContext('state'));
 
     let payload = $state.raw({});
     let expanded = $derived(resource.expanded);
+    let title = $derived(buttonText[expanded ? 'collapse' : 'expand']);
 
     $effect.pre(() => {
         payload = getPayload(resource);
     });
 
-    function onclick() {
+    function onclick(jsEvent) {
         resource.expanded = expanded = !expanded;
         toggle(payload.children);
         resources.length = resources.length;
         if (isFunction(resourceExpand)) {
-            resourceExpand({resource, view: toViewWithLocalDates(view)});
+            resourceExpand({resource, jsEvent, view: toViewWithLocalDates(view)});
         }
     }
 
@@ -39,8 +40,13 @@
 
 <span class="{theme.expander}">
     {#if payload.children?.length}
-        <button class="{theme.button}" {onclick}>
-            {#if expanded}&minus;{:else}&plus;{/if}
+        <button
+            class="{theme.button}"
+            aria-label="{title}"
+            title="{title}"
+            {onclick}
+            {@attach contentFrom(icons[expanded ? 'collapse' : 'expand'])}
+        >
         </button>
     {/if}
 </span>