+++ packages/core/src/plugins/resource-timeline/index.js
@@ -1,4 +1,4 @@
-import {btnTextDay, btnTextMonth, btnTextWeek, getPayload, themeView} from '#lib';
+import {assign, btnTextDay, btnTextMonth, btnTextWeek, getPayload, themeView} from '#lib';
 import {setExtensions} from '../time-grid/lib.js';
 import {createTRROptions, createTRRParsers} from '../time-grid/options.js';
 import {createRROptions} from '../resource-time-grid/options.js';
@@ -8,45 +8,59 @@ export default {
 	createOptions(options) {
         createTRROptions(options);
         createRROptions(options);
-		options.resourceExpand = undefined;
-        options.slotWidth = 32;
-		// Common options
-		options.buttonText.resourceTimelineDay = 'timeline';
-		options.buttonText.resourceTimelineWeek = 'timeline';
-		options.buttonText.resourceTimelineMonth = 'timeline';
-		options.theme.expander = 'ec-expander';
-        options.theme.rowHead = 'ec-row-head';
-		options.theme.slots = 'ec-slots';
-		options.view = 'resourceTimelineWeek';
-		options.views.resourceTimelineDay = {
-			buttonText: btnTextDay,
-			component: initViewComponent,
-			displayEventEnd: false,
-			dayHeaderFormat: {weekday: 'long'},
-			duration: {days: 1},
-			theme: themeView('ec-resource ec-timeline ec-day-view'),
-			titleFormat: {year: 'numeric', month: 'long', day: 'numeric'}
-		};
-		options.views.resourceTimelineWeek = {
-			buttonText: btnTextWeek,
-			component: initViewComponent,
-			displayEventEnd: false,
-			duration: {weeks: 1},
-			theme: themeView('ec-resource ec-timeline ec-week-view')
-		};
-		options.views.resourceTimelineMonth = {
-			buttonText: btnTextMonth,
-			component: initMonthViewComponent,
-			displayEventEnd: false,
-			dayHeaderFormat: {
-				weekday: 'short',
-				day: 'numeric'
+		assign(options, {
+			resourceExpand: undefined,
+			slotWidth: 32,
+			// Common options
+			view: 'resourceTimelineWeek'
+		});
+		assign(options.buttonText, {
+			expand: 'Expand',
+			collapse: 'Collapse',
+			resourceTimelineDay:  'timeline',
+			resourceTimelineWeek: 'timeline',
+			resourceTimelineMonth: 'timeline'
+		});
+		assign(options.icons, {
+			collapse: {html: '&minus;'},
+			expand: {html: '&plus;'}
+		});
+		assign(options.theme, {
+			expander: 'ec-expander',
+			rowHead: 'ec-row-head',
+			slots: 'ec-slots'
+		});
+		assign(options.views, {
+			resourceTimelineDay: {
+				buttonText: btnTextDay,
+				component: initViewComponent,
+				displayEventEnd: false,
+				dayHeaderFormat: {weekday: 'long'},
+				duration: {days: 1},
+				theme: themeView('ec-resource ec-timeline ec-day-view'),
+				titleFormat: {year: 'numeric', month: 'long', day: 'numeric'}
 			},
-			duration: {months: 1},
-			slotDuration: {days: 1},
-			theme: themeView('ec-resource ec-timeline ec-month-view'),
-			titleFormat: {year: 'numeric', month: 'long'}
-		};
+			resourceTimelineWeek: {
+				buttonText: btnTextWeek,
+				component: initViewComponent,
+				displayEventEnd: false,
+				duration: {weeks: 1},
+				theme: themeView('ec-resource ec-timeline ec-week-view')
+			},
+			resourceTimelineMonth: {
+				buttonText: btnTextMonth,
+				component: initMonthViewComponent,
+				displayEventEnd: false,
+				dayHeaderFormat: {
+					weekday: 'short',
+					day: 'numeric'
+				},
+				duration: {months: 1},
+				slotDuration: {days: 1},
+				theme: themeView('ec-resource ec-timeline ec-month-view'),
+				titleFormat: {year: 'numeric', month: 'long'}
+			}
+		});
 	},
 
     createParsers(parsers) {